+++ src/Examples/MultiProjectApp/MultiProjectApp.ConsoleApp/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-using NexusLabs.Needlr.Generators;
-
-[assembly: GenerateTypeRegistry(IncludeNamespacePrefixes = ["MultiProjectApp.ConsoleApp"])]