+++ src/Examples/SourceGen/AotSourceGenApp/GeneratorAssemblyInfo.cs
@@ -1,4 +1,0 @@
-using NexusLabs.Needlr;
-using NexusLabs.Needlr.Generators;
-
-[assembly: GenerateTypeRegistry(IncludeNamespacePrefixes = ["AotSourceGen", "AotSourceGenPlugin"])]